Remove Exodus data tie-in from post-deploy
The Vault kit forced our hand on this one; we don't get Exodus data
until we get past the post-deploy hook, since the Vault kit needs to use
the post-deploy hook to unseal the vault first.

`genesis info` has access to Exodus, so we'll just remove the env vars
listing from the post-deploy hook and leave it in the info hook.